+++ src/lib/dglib/include/DgBoundedRFBase.h
@@ -40,29 +40,29 @@ template<class B, class DB> class DgBoundedRFBase {
       const DgLocation& last  (void) const { return last_; }
       const DgLocation& end   (void) const { return end_; }
 
-      unsigned long long int size (void) const { return size_; }
+      uint64_t size (void) const { return size_; }
       bool validSize (void) const { return validSize_; }
 
       bool zeroBased (void) const { return zeroBased_; }
       void setZeroBased (bool zBasedIn) { zeroBased_ = zBasedIn; }
 
-      virtual unsigned long long int seqNum (const DgLocation& loc,
+      virtual uint64_t seqNum (const DgLocation& loc,
                                         bool convert = true) const = 0; 
 
       virtual bool lessThan (const DgLocation& loc1, 
                      const DgLocation& loc2, bool convert = true) const
                  { return (seqNum(loc1, convert) < seqNum(loc2, convert)); }
 
-      virtual DgLocation* locFromSeqNum (unsigned long long int sNum) const = 0;
+      virtual DgLocation* locFromSeqNum (uint64_t sNum) const = 0;
       
       // provide a generic interface to the discrete grid functionality
 
       virtual const DgRF<B, DB>& backFrame (void) const = 0;
       
-      virtual string dist2str (const long long int& dist) const = 0;
-      virtual long double dist2dbl (const long long int& dist) const = 0;
+      virtual string dist2str (const int64_t& dist) const = 0;
+      virtual long double dist2dbl (const int64_t& dist) const = 0;
 
-      virtual unsigned long long int dist2int (const long long int& dist) const = 0;
+      virtual uint64_t dist2int (const int64_t& dist) const = 0;
 
       virtual void setPoint (const DgLocation& loc, DgLocation& point) 
                                                                      const = 0;
@@ -94,7 +94,7 @@ template<class B, class DB> class DgBoundedRFBase {
 	   first_ (firstIn), last_ (lastIn), end_ (endIn),
 	   zeroBased_ (zBasedIn) {}
          
-      unsigned long long int size_;
+      uint64_t size_;
       bool validSize_;
 
    private:
