+++ src/lib/dglib/include/DgGeoProjRF.h
@@ -101,8 +101,8 @@ class DgGeoProjRF : public DgGeoDatumRF<DgDVec2D, long double> {
       virtual long double dist2dbl (const long double& dist) const
                        { return dist; }
 
-      virtual unsigned long long int dist2int (const long double& dist) const
-                       { return static_cast<unsigned long long int>(dist); }
+      virtual uint64_t dist2int (const long double& dist) const
+                       { return static_cast<uint64_t>(dist); }
 
       // the following routines are "back-doors" included for speed;
       // use with care!
