+++ src/lib/dglib/include/DgIDGG.h
@@ -173,8 +173,8 @@ class DgQ2DDRF : public DgRF<DgQ2DDCoord, long double> {
       virtual long double dist2dbl (const long double& dist) const
                        { return dist; }
 
-      virtual unsigned long long int dist2int (const long double& dist) const
-                       { return (unsigned long long int) dist; }
+      virtual uint64_t dist2int (const long double& dist) const
+                       { return (uint64_t) dist; }
 
       virtual const DgQ2DDCoord& undefAddress (void) const
                        { static DgQ2DDCoord undef; return undef; }
@@ -338,8 +338,8 @@ class DgVertex2DDRF : public DgRF<DgVertex2DDCoord, long double> {
       virtual long double dist2dbl (const long double& dist) const
                        { return dist; }
 
-      virtual unsigned long long int dist2int (const long double& dist) const
-                       { return (unsigned long long int) dist; }
+      virtual uint64_t dist2int (const long double& dist) const
+                       { return (uint64_t) dist; }
 
       static const DgVertTriVals& vertTable (int vertNum, int subTri) 
                        { return vertTable_[vertNum][subTri]; }
@@ -356,11 +356,11 @@ class DgVertex2DDRF : public DgRF<DgVertex2DDCoord, long double> {
 
 ////////////////////////////////////////////////////////////////////////////////
 class DgQ2DItoDConverter : 
-        public DgConverter<DgQ2DICoord, long long int, DgQ2DDCoord, long double>
+        public DgConverter<DgQ2DICoord, int64_t, DgQ2DDCoord, long double>
 {
    public:
 
-      DgQ2DItoDConverter (const DgRF<DgQ2DICoord, long long int>& from,
+      DgQ2DItoDConverter (const DgRF<DgQ2DICoord, int64_t>& from,
                           const DgRF<DgQ2DDCoord, long double>& to);
 
       const DgIDGG& IDGG (void) const { return *pIDGG_; }
@@ -376,12 +376,12 @@ class DgQ2DItoDConverter :
 
 ////////////////////////////////////////////////////////////////////////////////
 class DgQ2DDtoIConverter : 
-        public DgConverter<DgQ2DDCoord, long double, DgQ2DICoord, long long int>
+        public DgConverter<DgQ2DDCoord, long double, DgQ2DICoord, int64_t>
 {
    public:
 
       DgQ2DDtoIConverter (const DgRF<DgQ2DDCoord, long double>& from,
-                          const DgRF<DgQ2DICoord, long long int>& to);
+                          const DgRF<DgQ2DICoord, int64_t>& to);
 
       const DgIDGG& IDGG (void) const { return *pIDGG_; }
 
@@ -537,14 +537,14 @@ operator<< (ostream& stream, const DgInterleaveCoord& coord)
 { return stream << string(coord); }
 
 ////////////////////////////////////////////////////////////////////////////////
-class DgInterleaveRF : public DgRF<DgInterleaveCoord, long long int> {
+class DgInterleaveRF : public DgRF<DgInterleaveCoord, int64_t> {
 
    public:
 
       DgInterleaveRF (DgRFNetwork& networkIn, const string& nameIn)
-         : DgRF<DgInterleaveCoord, long long int>(networkIn, nameIn) { }
+         : DgRF<DgInterleaveCoord, int64_t>(networkIn, nameIn) { }
 
-      virtual long long int dist (const DgInterleaveCoord& add1, 
+      virtual int64_t dist (const DgInterleaveCoord& add1, 
                         const DgInterleaveCoord& add2) const 
                        { return 0; }
 
@@ -558,13 +558,13 @@ class DgInterleaveRF : public DgRF<DgInterleaveCoord, long long int> {
       virtual const char* str2add (DgInterleaveCoord* add, const char* str, 
                                    char delimiter) const;
 
-      virtual string dist2str (const long long int& dist) const
+      virtual string dist2str (const int64_t& dist) const
                        { return dgg::util::to_string(dist); }
 
-      virtual long double dist2dbl (const long long int& dist) const
+      virtual long double dist2dbl (const int64_t& dist) const
                        { return dist; }
 
-      virtual unsigned long long int dist2int (const long long int& dist) const
+      virtual uint64_t dist2int (const int64_t& dist) const
                        { return dist; }
 
       virtual const DgInterleaveCoord& undefAddress (void) const
@@ -574,12 +574,12 @@ class DgInterleaveRF : public DgRF<DgInterleaveCoord, long long int> {
 
 ////////////////////////////////////////////////////////////////////////////////
 class DgQ2DItoInterleaveConverter : 
-        public DgConverter<DgQ2DICoord, long long int, DgInterleaveCoord, long long int>
+        public DgConverter<DgQ2DICoord, int64_t, DgInterleaveCoord, int64_t>
 {
    public:
 
-      DgQ2DItoInterleaveConverter (const DgRF<DgQ2DICoord, long long int>& from,
-                                   const DgRF<DgInterleaveCoord, long long int>& to);
+      DgQ2DItoInterleaveConverter (const DgRF<DgQ2DICoord, int64_t>& from,
+                                   const DgRF<DgInterleaveCoord, int64_t>& to);
 
       const DgIDGG& IDGG (void) const { return *pIDGG_; }
 
@@ -637,13 +637,13 @@ class DgGridStats {
          : nCells_ (0), cls_ (M_ZERO), cellDistKM_ (M_ZERO), 
            cellAreaKM_ (M_ZERO), precision_ (DEFAULT_PRECISION) { }
 
-      unsigned long long int nCells     (void) const { return nCells_; }
+      uint64_t nCells     (void) const { return nCells_; }
       long double            cls        (void) const { return cls_; }
       long double            cellDistKM (void) const { return cellDistKM_; }
       long double            cellAreaKM (void) const { return cellAreaKM_; }
       unsigned int           precision  (void) const { return precision_; }
    
-      void setNCells (long long int nCellsIn) { nCells_ = nCellsIn; }
+      void setNCells (int64_t nCellsIn) { nCells_ = nCellsIn; }
       void setCLS    (long double clsIn)             { cls_ = clsIn; }
       void setCellDistKM (long double cellDistKMin)  { cellDistKM_ = cellDistKMin; }
       void setCellAreaKM (long double cellAreaIn)    { cellAreaKM_ = cellAreaIn; }
@@ -651,7 +651,7 @@ class DgGridStats {
       
    private:
 
-      unsigned long long int nCells_;
+      uint64_t nCells_;
       long double cls_;
       long double cellDistKM_;
       long double cellAreaKM_;
@@ -711,10 +711,10 @@ class DgIDGG : public DgDiscRF<DgQ2DICoord, DgGeoCoord, long double> {
       int                numAp4      (void) const { return numAp4_; }
       int                res         (void) const { return res_; }
       int                adjRes      (void) const { return adjRes_; }
-      long long int maxI        (void) const { return maxI_; }
-      long long int maxJ        (void) const { return maxJ_; }
-      long long int mag         (void) const { return mag_; }
-      long long int maxD        (void) const { return maxD_; }
+      int64_t maxI        (void) const { return maxI_; }
+      int64_t maxJ        (void) const { return maxJ_; }
+      int64_t mag         (void) const { return mag_; }
+      int64_t maxD        (void) const { return maxD_; }
       int                radix       (void) const { return radix_; }
       const string&      gridTopo    (void) const { return gridTopo_; }
       const string&      projType    (void) const { return projType_; }
@@ -772,7 +772,7 @@ class DgIDGG : public DgDiscRF<DgQ2DICoord, DgGeoCoord, long double> {
       // we'll give dummy definitions for these since some of our IDGG's may
       // not have defined these yet
 
-      virtual long long int dist (const DgQ2DICoord& add1, 
+      virtual int64_t dist (const DgQ2DICoord& add1, 
                                        const DgQ2DICoord& add2) const
            {
               report(string("DgIDGG::dist() this method has not been defined "
@@ -830,11 +830,11 @@ class DgIDGG : public DgDiscRF<DgQ2DICoord, DgGeoCoord, long double> {
       bool isSuperfund_;
       int  sfRes_;
 
-      long long int mag_;
-      long long int maxD_;
+      int64_t mag_;
+      int64_t maxD_;
       int radix_;
-      long long int maxI_;
-      long long int maxJ_;
+      int64_t maxI_;
+      int64_t maxJ_;
 
       DgQ2DICoord firstAdd_;
       DgQ2DICoord lastAdd_;
