+++ src/lib/dglib/include/DgIVec2D.h
@@ -31,7 +31,7 @@ class DgIVec2D {
 
       static const DgIVec2D& undefDgIVec2D;
 
-      DgIVec2D (long long int i = 0, long long int j = 0) 
+      DgIVec2D (int64_t i = 0, int64_t j = 0) 
        : i_(i), j_(j)
       {}
 
@@ -44,14 +44,14 @@ class DgIVec2D {
          j_ (dgg::util::lrint(pt.y())) 
       {}
 
-      void setI (long long int i) { i_ = i; }
-      void setJ (long long int j) { j_ = j; }
+      void setI (int64_t i) { i_ = i; }
+      void setJ (int64_t j) { j_ = j; }
 
       long double distance (const DgIVec2D& pt) const
               { return (pt - *this).magnitude(); }
 
-      long long int i (void) const { return i_; }
-      long long int j (void) const { return j_; }
+      int64_t i (void) const { return i_; }
+      int64_t j (void) const { return j_; }
       
       long double magnitude (void) const 
               { return sqrt((long double) (i_ * i_ + j_ * j_)); }
@@ -95,8 +95,8 @@ class DgIVec2D {
 
    private:
 
-      long long int i_;
-      long long int j_;
+      int64_t i_;
+      int64_t j_;
 
 };
 
@@ -110,8 +110,8 @@ DgIVec2D::scale (long double xScaleFactor, long double yScaleFactor)
 //
 ////////////////////////////////////////////////////////////////////////////////
 {
-   i_ = (long long int) dgg::util::lrint(i_ * xScaleFactor);
-   j_ = (long long int) dgg::util::lrint(j_ * yScaleFactor);
+   i_ = (int64_t) dgg::util::lrint(i_ * xScaleFactor);
+   j_ = (int64_t) dgg::util::lrint(j_ * yScaleFactor);
 
    return *this;
 
@@ -130,11 +130,11 @@ DgIVec2D::rotate (long double degrees)
    long double rotAng = degrees * M_PI_180;
    long double cosAng = cos(rotAng);
    long double sinAng = sin(rotAng);
-   long long int i = i_;
-   long long int j = j_;
+   int64_t i = i_;
+   int64_t j = j_;
 
-   i_ = (long long int) dgg::util::lrint(i * cosAng - j * sinAng);
-   j_ = (long long int) dgg::util::lrint(i * sinAng + j * cosAng);
+   i_ = (int64_t) dgg::util::lrint(i * cosAng - j * sinAng);
+   j_ = (int64_t) dgg::util::lrint(i * sinAng + j * cosAng);
 
    return *this;
